refactor: move CJK font rules to domain
